+++ src/_internal/pipe/index.ts
@@ -34,14 +34,14 @@ export type Func1 = (arg: any) => any
 export type ArgType<F, Else = never> = F extends (arg: infer A) => any
   ? A
   : Else
-type AsChain<F extends [Func1, ...Func1[]], G extends Func1[] = Tail<F>> = {
+type AsChain<F extends Functions, G extends Func1[] = Tail<F>> = {
   [K in keyof F]: (arg: ArgType<F[K]>) => ArgType<Lookup<G, K, any>, any>
 }
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export type Last<T extends any[]> = T extends [...infer F, infer L] ? L : never
 export type LaxReturnType<F> = F extends (...args: any) => infer R ? R : never
 
-export type Functions = [Func1, ...Func1[]]
+export type Functions = Func1[]
 export type PipedFunctions<F extends Functions> = AsChain<F> & F
 export type PipeResult<F extends Functions> = LaxReturnType<Last<F>>
 export type PipeFunction<F extends Functions, I> = (