+++ src/catchYobta/index.ts
@@ -1,4 +1,8 @@
-import type { Functions, SyncRulesPipeYobta } from '../_internal/pipe/index.js'
+import type {
+  Functions,
+  PipedFunctions,
+  SyncRulesPipeYobta,
+} from '../_internal/pipe/index.js'
 import { pipe } from '../_internal/pipe/index.js'
 import type {
   SyncRule,
@@ -51,7 +55,7 @@ export const catchYobta: CatchYobta = (
   ...rules: SyncRule<any, any>[]
 ) => {
   return ruleYobta((input, context) => {
-    const next = rules.map(rule => rule(context)) as Functions
+    const next = rules.map(rule => rule(context)) as PipedFunctions<Functions>
     try {
       return pipe(...next)(input)
     } catch (error) {