+++ src/itemsYobta/index.ts
@@ -1,6 +1,7 @@
 import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
 import type {
   Functions,
+  PipedFunctions,
   PipeFactoryResult,
   SyncRulesPipeYobta,
 } from '../_internal/pipe/index.js'
@@ -43,7 +44,7 @@ export const itemsYobta: ItemsYobta = <R extends SyncRules>(
   ...rules: R
 ): SyncRule<any[], PipeFactoryResult<R>[]> => {
   return ruleYobta((input: any[], context) => {
-    const next = rules.map(rule => rule(context)) as Functions
+    const next = rules.map(rule => rule(context)) as PipedFunctions<Functions>
 
     return input.map((item, index) => {
       try {