+++ src/syncYobta/index.ts
@@ -2,6 +2,7 @@ import { createContext } from '../_internal/createContext/index.js'
 import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
 import type {
   Functions,
+  PipedFunctions,
   PipeFactoryResult,
   SyncRulesPipeYobta,
 } from '../_internal/pipe/index.js'
@@ -54,7 +55,9 @@ export const syncYobta: SyncYobtaFactory =
   (data: any) => {
     const context = createContext(data)
 
-    const validators = rules.map(next => next(context)) as Functions
+    const validators = rules.map(next =>
+      next(context),
+    ) as PipedFunctions<Functions>
 
     try {
       const result: PipeFactoryResult<R> = pipe(...validators)(data)