+++ src/yobta/index.ts
@@ -3,6 +3,7 @@ import { createContext } from '../_internal/createContext/index.js'
 import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
 import type {
   Functions,
+  PipedFunctions,
   PipeFactoryResult,
   SyncRulesPipeYobta,
 } from '../_internal/pipe/index.js'
@@ -58,7 +59,9 @@ export const yobta: YobtaFactory =
       },
     }
 
-    const validators = rules.map(next => next(context)) as Functions
+    const validators = rules.map(next =>
+      next(context),
+    ) as PipedFunctions<Functions>
 
     try {
       return pipe(...validators)(data)